refactor: componentizando o formulario de produtos
