+++ front/src/renderer/src/views/pages/Menu/components/Products/modals/NewProductModal/index.tsx
@@ -1,18 +1,11 @@
-import Button from '@renderer/views/components/Button';
-import { Input } from '@renderer/views/components/Input';
 import Modal from '@renderer/views/components/Modal';
 
 import NewIngredientModal from '../../../Ingredients/NewIngredientModal';
-import CategoriesSection from '../../CategoriesSection';
-import IngredientsSection from '../../IngredientsSection';
-import UploadImage from '../../UploadImage';
 
 import useNewProductModal from './useNewProductModal';
 
-import { NumericInput } from '@renderer/views/components/NumericInput';
-import { Controller } from 'react-hook-form';
-import { CgCloseO } from 'react-icons/cg';
-import { Container, Content, Footer, ImageSection, InputsContainer } from './styles';
+import ProductForm from '../../ProductForm';
+import { Container } from './styles';
 
 interface NewProductModalProps {
   open: boolean
@@ -21,20 +14,10 @@ interface NewProductModalProps {
 
 export default function NewProductModal({ open, onClose }: NewProductModalProps) {
   const {
-    errors,
     width,
-    control,
     isLoading,
-    selectedCategoryId,
-    selectedIngredientsIds,
-    previewImageUrl,
     openNewIngredientModal,
-    register,
-    handleSelectedCategory,
-    handleSelectedIngredients,
     handleSubmit,
-    handleAddUploadImage,
-    handleRemoveUploadImage,
     handleCloseNewIngredientModal,
     handleOpenNewIngredientModal,
   } = useNewProductModal(onClose);
@@ -51,79 +34,11 @@ export default function NewProductModal({ open, onClose }: NewProductModalProps)
       $maxWidth={width < 1024 ? '500px' : '928px'}
     >
       <Container>
-        <Content>
-          <ImageSection>
-            <header>
-              <h3>Imagem</h3>
-
-              {errors.image && (
-                <div className="error">
-                  <CgCloseO color="#F63131" />
-                  <span>{errors.image.message}</span>
-                </div>
-              )}
-            </header>
-
-            <UploadImage
-              onUpload={handleAddUploadImage}
-              onRemoveImg={handleRemoveUploadImage}
-              previewImageUrl={previewImageUrl}
-            />
-          </ImageSection>
-
-          <InputsContainer>
-            <Input
-              type="text"
-              placeholder="Nome do produto*"
-              {...register('name')}
-              $error={errors.name?.message}
-              maxLength={30}
-            />
-
-            <Input
-              type="text"
-              placeholder="Descrição"
-              maxLength={110}
-              {...register('description')}
-              $error={errors.description?.message}
-            />
-
-            <p>Máximo 110 caracteres</p>
-
-            <Controller
-              control={control}
-              name="price"
-              render={({ field: { onChange, value, name } }) => (
-                <NumericInput
-                  name={name}
-                  placeholder="Valor do produto*"
-                  $error={errors.price?.message}
-                  onInputChange={onChange}
-                  value={value}
-                  maxLength={10}
-                />
-              )}
-            />
-          </InputsContainer>
-
-          <CategoriesSection
-            $error={errors.categoryId?.message}
-            onSelect={handleSelectedCategory}
-            selectedCategoryId={selectedCategoryId}
-          />
-
-          <IngredientsSection
-            openNewIngredientModal={handleOpenNewIngredientModal}
-            onSelected={handleSelectedIngredients}
-            selectedIngredientsIds={selectedIngredientsIds}
-          />
-        </Content>
-
-        <Footer>
-          <Button onClick={handleSubmit} isLoading={isLoading}>
-            Adicionar produto
-          </Button>
-        </Footer>
+        <ProductForm
+          onSubmit={handleSubmit}
+          onOpenIngredientModal={handleOpenNewIngredientModal}
+          isLoading={isLoading}
+        />
       </Container>
     </Modal>
   );
